# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InvenTrack API",
        "description": "REST API untuk aplikasi manajemen inventori InvenTrack.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/sales-orders": {
            "get": {
                "tags": [
                    "Sales Orders"
                ],
                "summary": "Daftar sales order",
                "operationId": "6a84e2fb007b22ac8fd8ae0b33e551cd",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "shipped",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List sales order"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sales Orders"
                ],
                "summary": "Buat sales order baru",
                "operationId": "f0fb2d6e279662647fe613d8440bb4c3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "order_date",
                                    "status",
                                    "details"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-12"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending",
                                        "enum": [
                                            "pending",
                                            "shipped",
                                            "cancelled"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Catatan SO"
                                    },
                                    "details": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "unit_price": {
                                                    "type": "number",
                                                    "example": 150000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sales order berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal atau stok tidak cukup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales-orders/{sales_order}": {
            "get": {
                "tags": [
                    "Sales Orders"
                ],
                "summary": "Detail sales order",
                "operationId": "1a0c929defdcd22e1924794a4b7d6da8",
                "parameters": [
                    {
                        "name": "sales_order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail sales order"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Sales Orders"
                ],
                "summary": "Update sales order",
                "operationId": "d0d8a79b73707d8b897e09450162aa58",
                "parameters": [
                    {
                        "name": "sales_order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "order_date",
                                    "status",
                                    "details"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-04-12"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "shipped",
                                        "enum": [
                                            "pending",
                                            "shipped",
                                            "cancelled"
                                        ]
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Catatan SO"
                                    },
                                    "details": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "unit_price": {
                                                    "type": "number",
                                                    "example": 150000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sales order berhasil diupdate"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Sales Orders"
                ],
                "summary": "Hapus sales order",
                "operationId": "74700f7f346020ae7b764b83bdbb0593",
                "parameters": [
                    {
                        "name": "sales_order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sales order berhasil dihapus"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}